Coding-Assistant-main/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
ollama>=0.1.0
google-generativeai
python-dotenv>=1.0.0
